fix Place Service, Repository and Controller

+++ src/main/java/co/develhope/team2/meeplemeet_project_team2/controllers/PlaceController.java
@@ -1,5 +1,6 @@
 package co.develhope.team2.meeplemeet_project_team2.controllers;
 
+import co.develhope.team2.meeplemeet_project_team2.DTO.PlaceDTO;
 import co.develhope.team2.meeplemeet_project_team2.entities.Place;
 
 import co.develhope.team2.meeplemeet_project_team2.services.PlaceService;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.lang.annotation.Repeatable;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,36 +24,67 @@ public class PlaceController {
     @PostMapping("/create")
     public ResponseEntity<Place> createPlace(@RequestBody Place place) {
         Place newPlace = placeService.createAPlace(place);
-        return new ResponseEntity<>(newPlace, HttpStatus.CREATED);
+        return ResponseEntity.ok(newPlace);
     }
 
     @GetMapping("/search/list")
     public @ResponseBody ResponseEntity<List<Place>> placeList() {
-        List<Place> ListPlace = placeService.getListOfPlaces();
-        return new ResponseEntity<>(ListPlace, HttpStatus.OK);
+        List<Place> listPlace = placeService.getListOfPlaces();
+        return ResponseEntity.ok(listPlace);
+    }
+
+    @GetMapping("/search/currenttime")
+    public @ResponseBody ResponseEntity<List<Place>> placeTimeList() {
+        List<Place> listTimePlace = placeService.findOpenPlace();
+        return ResponseEntity.ok(listTimePlace);
     }
 
     @GetMapping("/search/{id}")
     public @ResponseBody ResponseEntity<Optional<Place>> searchPlace(@PathVariable Integer id) {
-        Optional<Place> place = placeService.getPlaceById(id);
-        return new ResponseEntity<>(place, HttpStatus.OK);
+        Optional<Place> idplace = placeService.getPlaceById(id);
+        if(idplace.isPresent()) {
+        return ResponseEntity.ok(idplace);
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @PutMapping("/update/{id}")
     public @ResponseBody ResponseEntity<Place> updatePlace(@PathVariable Integer id, @RequestBody @NotNull Place place) {
-        Place updatePlace = placeService.updatePlace(id, place);
-        return new ResponseEntity<>(updatePlace, HttpStatus.OK);
+        Optional<Place> idPlace = placeService.getPlaceById(id);
+        if(idPlace.isPresent()) {
+            Place updatePlace = placeService.updatePlace(id, place);
+        return ResponseEntity.ok(updatePlace);
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    @DeleteMapping("delete/l/{id}")
+    public @ResponseBody ResponseEntity<Place> deleteLogicalPlace(@PathVariable Integer id) {
+        Optional<Place> idPlace = placeService.getPlaceById(id);
+        if(idPlace.isPresent()) {
+            placeService.deleteLogicalPlace(id, idPlace.get());
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @DeleteMapping("/delete/{id}")
     public @ResponseBody ResponseEntity<Place> deletePlace(@PathVariable Integer id) {
+        Optional<Place> idPlace = placeService.getPlaceById(id);
+        if(idPlace.isPresent()){
         placeService.deletePlaceById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @DeleteMapping("/delete/all")
     public @ResponseBody ResponseEntity<Place> deleteAllPlace() {
         placeService.deleteAllPlace();
-        return new ResponseEntity<>(HttpStatus.OK);
+        return ResponseEntity.ok().build();
     }
 }
